backend/ocr_llm/requirements.txt
paddlepaddle
paddleocr
google-generativeai
numpy
opencv-python
python-dotenv
